+++ R/theme_grid.R
@@ -12,8 +12,8 @@ theme_grid <- function () {
             legend.key = element_rect(fill="transparent", colour=NA),
             legend.title=element_text(size=14),
             legend.text=element_text(size=12),
-            axis.line = element_line(color = "black", size = .75, linetype = "solid"),
-            axis.ticks = element_line(colour = "black", size = .75),
+            axis.line = element_line(color = "black", linewidth = .75, linetype = "solid"),
+            axis.ticks = element_line(colour = "black", linewidth = .75),
             panel.background  = element_blank(),
             panel.grid.minor = element_blank(),
             panel.grid.major = element_line(color="lightgrey"),
@@ -25,7 +25,7 @@ theme_grid <- function () {
             axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), angle = 90, size = 14),
             axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), angle = 0, size = 14),
             axis.ticks.length = unit(0.20, "cm"),
-            strip.background = element_rect(color="black", size=.75, linetype="solid"),
+            strip.background = element_rect(color="black", linewidth=.75, linetype="solid"),
             strip.text.x = element_text(size = 12, color = "black"),
             strip.text.y = element_text(size = 12, color = "black")
         )